requirements.txt
aniso8601==3.0.0
click==6.7
configparser==3.5.0
Flask==0.12.2
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
paho-mqtt==1.3.1
pytz==2018.4
six==1.11.0
Werkzeug==0.14.1
